+++ crates/facade/src/use_case.rs
@@ -1,15 +1,25 @@
-use qsimplify::{dto::GateOperation, dto::mapper, simplifier};
+use qsimplify::{
+    DiracFormat, PiFormat,
+    dto::{GateOperation, mapper},
+    simplifier,
+};
 
 #[must_use]
 pub fn display_graph(operations: &[GateOperation]) -> String {
     let graph = mapper::operations_to_graph(operations);
-    graph.to_string()
+    graph.display_nodes_and_edges(PiFormat::Fancy)
 }
 
 #[must_use]
 pub fn display_grid(operations: &[GateOperation]) -> String {
     let graph = mapper::operations_to_graph(operations);
-    graph.display_grid()
+    graph.display_grid(PiFormat::Fancy)
+}
+
+#[must_use]
+pub fn display_matrix(operations: &[GateOperation]) -> String {
+    let graph = mapper::operations_to_graph(operations);
+    graph.display_matrix(DiracFormat::Fancy)
 }
 
 #[must_use]