+++ crates/library/src/interface/display/mod.rs
@@ -1,8 +1,11 @@
 pub(crate) mod angle_formatter;
 pub(crate) mod graph_display;
+pub(crate) mod number_formatter;
 pub(crate) mod projection_display;
 
 #[cfg(test)]
 mod angle_formatter_tests;
 #[cfg(test)]
 mod graph_display_tests;
+#[cfg(test)]
+mod number_formatter_tests;